Create move_set directory
This directory will contain all of the valid moves for a chess piece.

**How it will work**
When a player selects an initial position for a piece, the corresponding
file will find all of the "valid moves" of that piece.

The player will then need to input a position to where they want to move the piece.
If the selected "new position" for the piece is not in "valid moves", then
an error will be thrown and the player will have to select a different
"new position".

Valid moves will most likely be stored via array